Fixed the error that stopped images from displaying
When Jekyll is used offline i.e. hosted at 127.0.0.1:4000 (localhost via bundle) The root level of the generated site is in the same level as the index.md file. However when hosted in github pages it seems that each .md file becomes a root to the relative locations it's referent to. therefore to point to the global location I used the ".." operator to reach the same level as the index.md file.